Also support constant expressions on Rust < 1.32
On Rust 1.32 we also support constant expressions now,
with the cost that we are never using concat!,
always format!.
This might bloat the generated binary a bit as stuff like
format!("v {}", 11) usually isn't const elided by llvm.
But ultimately the effects are small and I expect
that convenience is more important to most users
than this small difference in binary size.
If someone with tight size constraints wants to use
this library with Rust < 1.32, bug reports are welcome.
Then this descision might be revisited.